Feat/authentication

+++ src/main/java/ro/unibuc/hello/service/UserService.java
@@ -3,45 +3,77 @@ package ro.unibuc.hello.service;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 
 import jakarta.persistence.EntityNotFoundException;
 import lombok.AllArgsConstructor;
 import ro.unibuc.hello.data.UserRepository;
+import ro.unibuc.hello.data.Role;
 import ro.unibuc.hello.data.UserEntity;
 import ro.unibuc.hello.dto.request.RegisterDto;
 import ro.unibuc.hello.dto.response.UserDto;
 import ro.unibuc.hello.dto.response.UserListDto;
 import ro.unibuc.hello.exception.EntityAlreadyExistsException;
+
+import java.util.List;
 import java.util.Optional;
 
-@Component
+@Service
 @AllArgsConstructor
-public class UserService {
+public class UserService implements UserDetailsService{
     
     @Autowired
     private UserRepository userRepository;
-
     private ModelMapper modelMapper;
 
-    public UserDto createUser(RegisterDto registerDto){
-        userRepository.findByUsername(registerDto.getUsername())
-                .ifPresent(user -> { throw new EntityAlreadyExistsException(); });
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+            UserEntity user = userRepository.findByUsername(username)
+                .orElseThrow(()->new EntityNotFoundException(username));
 
-        var student = modelMapper.map(registerDto, UserEntity.class);
+        GrantedAuthority authority = new SimpleGrantedAuthority("ROLE_" + user.getRole());
 
-        var savedEntity = userRepository.save(student);
-
-        return modelMapper.map(savedEntity, UserDto.class);
+        return new org.springframework.security.core.userdetails.User(
+                user.getUsername(),
+                user.getPassword(),
+                List.of(authority)
+        );
     }
 
-    public UserEntity loadUser(String username) {
+    // Helper method: get any user from database using their username
+    private UserEntity loadUser(String username) {
         return userRepository.findByUsername(username)
-                                .orElseThrow(()->new EntityNotFoundException(username));
+                                .orElseThrow(()->new EntityNotFoundException("user"));
     }
+
+    //Helper method: get the logged in user
+
+    public UserEntity getAuthenticatedUser(){
+        String username = ((UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUsername();
+        return userRepository.findByUsername(username)
+                                .orElseThrow(()->new EntityNotFoundException("user"));
+    }
+
+    public UserDto getSelf(){
+        var user = getAuthenticatedUser();
+        return modelMapper.map(user,UserDto.class);
+    }
+    
+
     public UserDto getUser(String username) {
         var user = loadUser(username);
         return modelMapper.map(user, UserDto.class);
     }
+
 
     public UserDto updateUser(String username, RegisterDto registerDto){
         var user = userRepository.findByUsername(username)
@@ -69,6 +101,7 @@ public class UserService {
 
     public void delete(String username){
         var user = loadUser(username);
-        userRepository.delete(user);
+        if (username!=getAuthenticatedUser().getUsername() && getAuthenticatedUser().getRole()!=Role.ADMIN)
+            userRepository.delete(user);
     }
 }
